+++ examples/01_counter/src/index.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import App from './App';
+import App from './app';
 
 const ele = document.getElementById('app');
 if (ele) {
